Updating LWIP related code
It seems there is no ChibiOS function available to delay during board
init, so the simple loop delay counter will have to stay. Changing to
use a volatile counter in the loop as if optimised out at -02 it will
hang in mac_lld_start.

Adding in some performance related tweaks.